requirements.txt
altair
numpy
pandas
pydeck
streamlit
google-generativeai
pdf2image
beautifulsoup4
PyPDF2